refactor(drivers): extract shared acp_native base from gemini/kimi/opencode (#111)
Three ACP-native drivers (gemini, kimi, opencode) each carried ~1500-2700
lines of structurally near-identical code: reader loop, response routing,
session lifecycle, cancel/close, ensure_started semantics, EOF drain,
permission auto-approval. Bug fixes had to be applied in three places and
behaviors drifted apart at edges.

Move all of it into `src/agent/drivers/acp_native/`:
- `mod.rs` — `AcpDriverConfig` (struct of fn pointers + bools + enums) +
  `InitPromptStrategy` + shared `open_session` helper.
- `state.rs` — `SharedReaderState`, `PendingRequest`, `SessionState`.
- `core.rs` — `AcpNativeCore`, `ensure_started` (race-safe lazy spawn,
  non-sticky failure), `spawn_and_initialize`, `is_stale`, `Drop`.
- `handle.rs` — `AcpNativeHandle` + full `Session` impl: run, prompt,
  cancel, close (with `closed_emitted` race guard).
- `reader.rs` — `reader_loop`, `handle_response`, `handle_session_update`,
  `pick_session*`. Routes responses by JSON-RPC id through
  `pending_requests` map (avoids `acp_protocol::parse_line`'s id-bucketing
  that misclassifies `session/new` at id≥3 as PromptResponse).
- `tests.rs` — 19 generic tests with a `TestConfig`. Audit table at top
  maps each pre-migration per-driver test to its shared equivalent.

Per-runtime variation lives entirely in the static `&'static AcpDriverConfig`
each driver owns. No trait, no generics — three runtimes, single
instantiation, function-pointer dispatch.

Behavior preserved bit-for-bit:
- Cancel stays local-only.
- `stopReason` continues to be ignored; all completions emit Natural.
- `session/close` remains local-only (no RPC).
- No capability checking before `session/load`; no `session/resume`.
- HTTP MCP transport stays as-is.
The 8 ACP spec gaps catalogued in the plan are tracked as follow-up issues
— each becomes a 1-place fix now that the base is shared.

New shared test `ensure_started_concurrent` closes a coverage gap: drives
two concurrent `ensure_started` calls and asserts the slow path runs
exactly twice (each caller retries after its predecessor fails, proving
serialization + non-stickiness without needing a real runtime binary).

Opencode shape conversion: collapse the `FactoryPath::Bootstrap | Secondary`
split into the unified handle. The race the bootstrap protected against
(deferred prompt id colliding with a racing secondary `new_session`)
cannot occur in the unified model — `ensure_started` serializes through
`start_in_progress`, and `alloc_id` runs only after that mutex is
released. Deletes `OpencodeAgentProcess`, `FactoryPath`,
`run_bootstrap*`, `send_deferred_bootstrap_prompt`, the local
`classify_line`/`dispatch_line`, and the bootstrap-only state fields
(`bootstrap_pending_prompt`, `bootstrap_session_id`,
`bootstrap_requested_session_id`).

Diff:
- gemini.rs   1718 → 423 (−75%)
- kimi.rs     2737 → 328 (−88%)
- opencode.rs 2834 → 339 (−88%)
- net: 7289 → 3650 lines (−50% across drivers + new shared module incl. tests)

Verified: cargo test (527 passed), cargo test --test e2e_tests (10 passed),
cargo clippy --lib --tests -- -D warnings (clean).

Plan: docs/plans/2026-04-27-acp-native-driver-unification-plan.md
